login and logout completed
current user is being passed an obejct to  subcomponents via props. The object is nested inside another object with the same name. so in the sub component to get the current users email, we need to use currentUser.currentUser.email, This is unintended and should be fixed in the future.